requirements.txt
streamlit==1.23.1
protobuf==3.20.0
mysql-connector==2.2.9